a2/requirements.txt
pyquaternion
PyOpenGL
PyOpenGL_accelerate
